# src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/posts": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "searchMode",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/posts/{postId}": {
      "put": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "categorySpace": {
                  "example": "any"
                },
                "categoryInterest": {
                  "example": "any"
                },
                "categoryStudyMate": {
                  "example": "any"
                },
                "textContent": {
                  "example": "any"
                },
                "youtubeUrl": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "505": {
            "description": "HTTP Version Not Supported"
          }
        }
      },
      "delete": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
...
}